Sprint to Master

Closes OP-157 and OP-200

See merge request open-platform/api!192

+++ src/main/kotlin/io/openfuture/api/domain/event/FundsDepositedEvent.kt
@@ -2,10 +2,10 @@ package io.openfuture.api.domain.event
 
 import io.openfuture.api.component.web3.event.EventType
 import io.openfuture.api.component.web3.event.EventType.FUNDS_DEPOSITED
-import java.math.BigInteger
+import java.math.BigDecimal
 
 data class FundsDepositedEvent(
-        val amount: BigInteger,
+        val amount: BigDecimal,
         val toAddress: String
 ) : Event {
